Deleting functions

+++ Chapter 10/FunctionOverload/main.cpp
@@ -4,7 +4,7 @@ int add(int x, int y) {
     return x+y;
 }
 
-int add(int x, int y) {
+float add(float x, float y) {
     return x+y;
 }
